Fixed Trac bug 65: Doesn't show correct position after 203:50 mins - caused by an unsigned long overflow.

+++ Audio/Chain/OutputNode.h
@@ -18,7 +18,7 @@
 @interface OutputNode : Node {
 	AudioStreamBasicDescription format;
 	
-	unsigned long amountPlayed;
+	unsigned long long amountPlayed;
 	OutputCoreAudio *output;
 }
